+++ bld/plustest/regress/cool/examples/ex4_9.c
@@ -25,8 +25,8 @@ int main (void) {
   for (long i = 0; i < 1000000L; i++)   // Loop for 1000000 times and
     sum = sum + i;                                      // Sum up numbers
   long elapsed = t1.real() + 1; // make sure it's non-zero
-  elapsed = log10(elapsed) + 1;
-  elapsed = log10(elapsed);
+  elapsed = log10(double(elapsed)) + 1;
+  elapsed = log10(double(elapsed));
   cout << "Summation of integers from 0 through 1000000 (" << sum << "=" << check << ") took ";
   cout << "log10(log10(time in ms)) = " << elapsed << endl;
   return 0;                                     // Return valid competion code
